backend/requirements.txt
fastapi>=0.103.0
uvicorn[standard]>=0.23.2
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.7
pydantic>=2.4.0
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
alembic>=1.12.0
PyMuPDF>=1.23.0
python-multipart
huggingface_hub
redis
passlib[bcrypt]
python-jose[cryptography]
sentence-transformers
faiss-cpu
numpy
pgvector
pytest
pytest-asyncio
pytest-mock
